+++ src/Data/Event/BeforeObjectEventInterface.php
@@ -1,8 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Remorhaz\JSON\Data\Event;
-
-interface BeforeObjectEventInterface extends BeforeStructureEventInterface
-{
-}