+++ src/Data/Value/DecodedJson/NodeScalarValue.php
@@ -3,9 +3,6 @@ declare(strict_types=1);
 
 namespace Remorhaz\JSON\Data\Value\DecodedJson;
 
-use Generator;
-use Iterator;
-use Remorhaz\JSON\Data\Event\ScalarEvent;
 use Remorhaz\JSON\Data\Path\PathInterface;
 use Remorhaz\JSON\Data\Value\NodeValueInterface;
 use Remorhaz\JSON\Data\Value\ScalarValueInterface;
@@ -34,15 +31,5 @@ final class NodeScalarValue implements NodeValueInterface, ScalarValueInterface
     public function getPath(): PathInterface
     {
         return $this->path;
-    }
-
-    public function createEventIterator(): Iterator
-    {
-        return $this->createGenerator();
-    }
-
-    private function createGenerator(): Generator
-    {
-        yield new ScalarEvent($this);
     }
 }