+++ src/Value/EvaluatedValue.php
@@ -3,10 +3,6 @@ declare(strict_types=1);
 
 namespace Remorhaz\JSON\Path\Value;
 
-use Generator;
-use Iterator;
-use Remorhaz\JSON\Data\Event\ScalarEvent;
-
 final class EvaluatedValue implements EvaluatedValueInterface
 {
 
@@ -20,15 +16,5 @@ final class EvaluatedValue implements EvaluatedValueInterface
     public function getData(): bool
     {
         return $this->value;
-    }
-
-    public function createEventIterator(): Iterator
-    {
-        return $this->createGenerator();
-    }
-
-    private function createGenerator(): Generator
-    {
-        yield new ScalarEvent($this);
     }
 }